Use message id from mbox header if none exists in MIME.
Some messages (like chats) don't have a Message-Id mime header, so the message is saved without a primary key.

A previous commit used the thread id in this situation, but the same thread id can be used for multiple messages. This id, which is the message id used by the gmail api, should be unique across all messages.